+++ src/v2/conversions.rs
@@ -2047,7 +2047,10 @@ impl TryFrom<baker_event::Event> for super::types::BakerEvent {
                 data: Box::new(v.try_into()?),
             },
             baker_event::Event::BakerSetOpenStatus(v) => {
-                let open_status = v.open_status().into();
+                let open_status = match v.open_status() {
+                    s if s <= OpenStatus::ClosedForAll => Upward::Known(s.into()),
+                    _ => Upward::Unknown,
+                };
                 Self::BakerSetOpenStatus {
                     baker_id: v.baker_id.require()?.into(),
                     open_status,